refactor : 인기 상품 조회와 상품 조회 분리
