+++ src/main/java/kr/hhplus/be/server/interfaces/api/product/ProductController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import kr.hhplus.be.server.interfaces.api.ApiResponse;
-import kr.hhplus.be.server.interfaces.api.product.response.BestSellerReadAllResponse;
-import kr.hhplus.be.server.interfaces.api.product.response.BestSellerSimpleInfo;
+import kr.hhplus.be.server.interfaces.api.bestseller.response.BestSellerReadAllResponse;
+import kr.hhplus.be.server.interfaces.api.bestseller.response.BestSellerSimpleInfo;
 import kr.hhplus.be.server.interfaces.api.product.response.ProductReadAllResponse;
 import kr.hhplus.be.server.interfaces.api.product.response.ProductSimpleInfo;
 
@@ -27,17 +27,5 @@ public class ProductController implements ProductControllerSpec {
 		);
 		return ApiResponse.ok(response);
 
-	}
-
-	@GetMapping("/best")
-	public ApiResponse<BestSellerReadAllResponse> getBestSellerLimitFive() {
-		BestSellerReadAllResponse response = new BestSellerReadAllResponse(
-			List.of(
-				new BestSellerSimpleInfo(1L, "상품1", 100L, 10L, BigDecimal.valueOf(10000)),
-				new BestSellerSimpleInfo(2L, "상품2", 200L, 20L, BigDecimal.valueOf(20000))
-			)
-		);
-
-		return ApiResponse.ok(response);
 	}
 }
